pipeline/requirements.txt
GitPython==3.1.43
GitPython==3.1.44
langchain==0.3.18
langchain_core==0.3.34
langchain_google_genai==2.0.9
langchain_openai==0.3.4
langgraph==0.2.70
protobuf==5.29.3
PyGithub==2.5.0
PyGithub==2.5.0
Requests==2.32.3
